Fix unitialized name buffer in CPIO parser
Fixes a possible stack buffer overflow introduced in 0.103 development
when we added optional names to file maps (fmaps). The CPIO parser uses
a stack buffer to store the name (if present).  If no name present, then
the stack buffer was passed unitialized to the fmap scanning function
which could cause an overflow.

This fix both initializes the buffer and uses a pointer so the scan
function gets NULL instead of a buffer in the event that a name isn't
present as that's the intended way to use the API, rather than passing
an empty string name buffer.