configure.ac: add workaround to support @runstatedir@.
This commit adds a workaround to configure.ac to support the many
distributions who patch "runstatedir" support into their autoconf.
The point of "runstatedir" variable is to make the location of the
runtime state directory configurable independent of "localstatedir",
so that, for example, PID files can be placed in /run instead of the
traditional /var/run.

Autoconf v2.70 supports runstatedir out of the box, but it has not
been released yet. When it is, and after AC_PREREQ has been bumped,
this chunk of code can simply be deleted. In the meantime it lets
us replace @runstatedir@ within certain files by running e.g.

  $ make runstatedir=/run

to override the default at build time. The default value is
@localstatedir@/run, and is backwards-compatible.