clamav-milter,clamd,clamonacc,freshclam: add OpenRC service scripts.
This commit adds four new OpenRC service scripts under <daemon>/openrc:

  * clamav-milter
  * clamd
  * clamonacc
  * freshclam

These scripts are primarily useful on Gentoo, where OpenRC is the
default service manager, and on Debian (and derivatives) where it is a
supported option. The scripts belong upstream because they need to
know a few paths that are provided to the build system, namely the
place where the executables themselves are located, and the place
where PID files are stored. So, the (s)bindir and runstatedir.

A forthcoming commit will perform substitutions for those directory
variables at build-time, eliminating the need for every distribution
to customize its own service scripts.