etc: convert conf files to templates and make @VARIABLE@ substitutions.
The sample configuration files for clamav-milter, clamd, and freshclam
all have PidFile directives that can make use of the @RUNSTATEDIR@
variable, since one primary purpose of the runstatedir is to hold PID
files. Particularly clamav-milter and clamd must keep the
configuration file directive and the "pidfile" service script
variables synchronized. This commit inserts @RUNSTATEDIR@ into those
commented directives, such that they agree with the corresponding
OpenRC service scripts. The PidFile directive in freshclam.conf is
ignored by its OpenRC service script -- as should be the case with the
others eventually -- but the same change was made there for consistency.

This commit also adds an "editgnudirs" command to the top-level
Makefile.am that is used to substitute @BINDIR@, @SBINDIR@, and
@RUNSTATEDIR@ into a file. Or more precisely, it substitutes those
variables into a template file named "foo.in" to produce "foo". That
new rule is then applied to the sample configuration files which have
been renamed to (for example) clamd.conf.sample.in.in. The ./configure
script processes that to clamd.conf.sample.in, and "make" applies
variable substitutions to produce clamd.conf.sample.

The end result of this change is that the OpenRC service scripts work
out-of-the-box when the clamd and clamav-milter PidFile directives are
merely uncommented, and do not need to be edited.