*/Makefile.am: allow @VARIABLE@ replacement in OpenRC scripts.
This commit adds the necessary magic to Makefile.am for clamav-milter,
clamd, clamonacc, and freshclam to convinces the build system to
install the corresponding OpenRC service scripts. For now, the scripts
are installed to $(sysconfdir)/init.d, which is generally the correct
location -- unless OpenRC itself was pointed at a non-standard path
The location could be made configurable (as with the systemd service
path), but at the moment that seems like overkill.