m4/reorganization/libs/pcre.m4: grab PCRE_CPPFLAGS from pkg-config, too.
The preprocessor flags (and in particular, the -I flags) are left out
of the commit that introduced pkg-config support for libpcre2. This
commit adds them. The obviously-incorrect addition of pkg-config's
PCRE_CFLAGS to the global PCRE_CPPFLAGS is in fact the correct thing
to do, because PCRE_CFLAGS (and pkg-config --cflags) is mis-named:
it contains pre-processor flags, as you can verify from the upstream
pcre2 *.pc files.

This fixes the build on FreeBSD, where the libpcre headers otherwise
cannot be found.