clamd: time-of-check/time-of-use issue
Clamd presently uses an access() check prior to a scan to see if the
process has permission to open the file. This is problematic for two
reasons.
1. There's a time-of-check/time-of-use issue with checking access
   before accessing because the permissions could change in between.
2. The POSIX CAP_DAC_READ_SEARCH feature won't work with access().

CAP_DAC_READ_SEARCH is a posix capability that allows the kernel to
bless a process, saying "You can read any file just like you were root."

The service manager can run clamd and bless it at run time with
CAP_DAC_READ_SEARCH such that fd-passing is not required for clamd to
read files it would otherwise not have permission to access.