bb12597: Fix daemon switch user issues
Drop privileges in the parent process before waiting for the signal 
from the child process, so that the parent properly responds to
the signal and terminates.

Verify that the log file will be owned by user that the deamon will
run as.

Explicitly set PID file ownership to root when starting the daemon 
as root with the PID file enabled.