Update valgrind suppression rule
The previous valgrind suppression rule for wcsrtombs assumed ascii
conversion.  With the clamd & clamscan change to use setlocale() to get
utf8 filenames to work in UnRAR, the wcsrtombs stack trace changed to
use utf8.

This patch supplements valgrind.supp with the new stack trace.  The old
one was left in, just in case.