freshclam autotools: Fix libshared.a linking issue
As libfreshclam links in libshared.a, we don't need to link it again
into freshclam. Presently we do and as a consequence  the global
variables in libshared.a cannot be reliably set in freshclam/libfreshlam
and be expected to have the correct values within libshared.so calls.

Specifically, users observed freshclam failing to use syslog even though
syslog was enabled. The reason was that the libshared.a logg() function
will only write to syslog() if the logg_syslog global is non-zero.
When libshared.a is linked twice (first into libfreshclam and again into
freshclam) then setting the logg_syslog global might not set the same
variable that the logg() function depends on.

To address this issue this patch stops linking libshared.a into
freshclam and instead exports the required libshared.a symbols from
libfreshclam.so (using libfreshclam.map) so that freshclam can still use
those functions.