etc/clamd: synchronize the (disabled) default LocalSocket directory.
When using a local socket, the service manager needs to create a
directory that is writable only by the clamav user. Typically
/run/clamav is used. The clamd daemon then creates a socket named, for
example, /run/clamav/clamd.ctl. That's currently what the clamd
systemd service does, and this commit brings the sample configuration
file and the OpenRC service to parity.

First, the default LocalSocket location was updated in the sample
configutation file to point to @RUNSTATEDIR@/clamav/clamd.ctl. The
previous path "/tmp/clamd.socket" is a security vulnerability, the
setting is disabled by default, and the new path is what everyone is
using anyway modulo perhaps the name of the socket. So the change is
relatively backwards-compatible.

Second, the clamd OpenRC service was updated to create the directory
@RUNSTATEDIR@/clamav and to make it writable by the clamav user/group.
This makes local socket configurations work almost out-of-the-box
(after you uncomment the setting), and is harmless otherwise.