CMake: Fix ncurses linking issue on FreeBSD
The linker is unable to find libncurses.so.6.2 on FreeBSD.
If I change Curses::curses from an UNKNOWN library to an
INTERFACE library, then it works fine. While UNKNOWN
appears to be preferred by the CMake community, it seems
like it doesn't add -L/usr/local/lib to the LDFLAGS on my
FreeBSD test machine where /usr/local isn't used by
default.

Disclaimer: This fix feels like a bit of a hack, but it works for me.