Reduce scan time for large PNG files
PNG file image data is decompressed to determine if it exceeds the
calculated image size as a heuristic for CVE-2010-1205. However, the
image data isn't actually scanned so we can reduce scan time and RAM
usage by not allocating space for the entire decompressed image.

Also skipped decompression of image data if the calculated image data
excceeds max-scansize.