clamonacc: Reduce warning log verbosity
Users have complained about two specific log events that are extremely
verbose in non-critical error conditions:

- clamonacc reports "ERROR: Can't send to clamd: Bad address"

  This may occur when small files are created/destroyed before they can
  be sent to be scanned. The log message probably should only be
  reported in verbose mode.

- clamonacc reports "ClamMisc: $/proc/XXX vanished before UIDs could be
  excluded; scanning anyway"

  This may occur when a process that accessed a file exits before
  clamonacc find out who accessed the file. This is a fairly frequent
  occurence. It can still be problematic if `clamd` was the process which
  accessed the file (like a clamd temp file if watching /tmp), generally
  it's not an issue and we want to silently scan it anyways.

Also addressed copypaste issue in onas_send_stream() wherein fd is set
to 0 (aka STDIN) if the provided fd == 0 (should've been -1 for invalid
FD) and if filename == NULL. In fact clamonacc never scans STDIN so the
scan should fail if filename == NULL and the provided FD is invalid
(-1).

I also found that "Access denied. ERROR" is easily provoked when using
--fdpass or --stream using this simple script:

  for i in {1..5000}; do echo "blah $i" > tmp-$i && rm tmp-$i; done

Clamdscan does not allow for scans to fail quietly because the file does
not exist, but for clamonacc it's a common thing and we don't want to
output an error. To solve this, I changed it so a return length of -1
will still result in an "internal error" message but return len 0
failures will be silently ignored.

I added a static variable to onas_client_scan() that keeps state in case
clamd is stopped and started - that way it won't print an error message
for every event when offline. Instead it will log an error for the first
connection failure, and log again when the connection is re-established
for a future scan. Calls to onas_client_scan() are already wrapped with
the onas_scan_lock mutex so the static variable should be safe.

Finally, there were a couple of error responses from clamd that can
occur if the file isn't found which we want to silently ignore, so I've
tweaked the code which checks for specific error messages to account for
these.