clamd: log the file name when using fd-passing
This improvement looks up the filename given the file descriptor.
This is supported on Mac and Linux but not presently supported
on other UNIX operating systems. FD-passing is not available on
Windows.

On supported systems, the verdict in the clamd log and the VirusEvent
will show the actual file path instead of something like fd[14].