bb12522: Fix freshclam db test in daemon mode
Database validation was broken when running daemon mode.
The source of this bug was a mistake when overhauling freshclam.
The freshclam overhaul accidentally removed the call to disable
waiting for child processes in the global signal handler.
This patch fixes the issue by disabling the wait prior to the
fork to test the database.