tests: limit # connections in test_connections
The test previously tried to limit the # of connections to
`ulimit -n` - 5.  On most linux docker containers this failed
with a test timeout at about connection 285 or so.
The output in test-clamd.log would look something like this:

check_clamd.c:74:E:clamd stress test:test_connections:0: (after this
point) Test timeout expired

The same issue was observed with FreeBSD (12.2) when limiting to around
280 (noting that the FD # in the debug log actually hit around 288).

Limiting the # of connections to 250 resolves the issue in our test
pipeline.